+++ code/mathematics/gcd.cpp
@@ -1,1 +1,1 @@
-int gcd(int a, int b) { return b == 0 ? a : gcd(b, a % b); }
+ll gcd(ll a, ll b) { return b == 0 ? a : gcd(b, a % b); }
