+++ code/mathematics/mod_inv.test.cpp
@@ -1,12 +1,12 @@
 #include "../code/mathematics/gcd.cpp"
 
 void test() {
-    for (int a = -20; a < 20; a++) {
-        for (int b = 2; b < 40; b++) {
+    for (ll a = -20; a < 20; a++) {
+        for (ll b = 2; b < 40; b++) {
             if (gcd(a, b) == 1) {
-                assert_equal(1, smod((a * mod_inv(a, b)), b));
+                assert_equal(1LL, smod((a * mod_inv(a, b)), b));
             } else {
-                assert_equal(-1, mod_inv(a, b));
+                assert_equal(-1LL, mod_inv(a, b));
             }
         }
     }
